JAT-0048 Refactor frontend infra scripts
