+++ liferay/teams/frontend_infra/frontend_infra_constants.py
@@ -1,0 +1,12 @@
+class FileName:
+    OUTPUT_MESSAGE_FILE_NAME = "output_message_fi.txt"
+    OUTPUT_INFO_FILE_NAME = "output_info_fi.txt"
+    OUTPUT_BUG_THRESHOLD_INFO_FILE_NAME = "bug_threshold_output_info_fi.txt"
+
+
+class Roles:
+    QA_JIRA_USER = 'carlos.brichete'
+
+
+class Sheets:
+    FI_TESTMAP_ID = '1_liLRC1XHBydH_mfeeDifgKxBWN3kfRtX7uqbKgJ72k'
